// src/types/KazagumoTypes.ts
export interface KazagumoTrack {
  track: string;
  info: {
    identifier: string;
    title: string;
    author: string;
    length: number;
    artworkUrl?: string;
    uri?: string;
    sourceName: string;
    isSeekable: boolean;
    isStream: boolean;
    position?: number;
  };
  pluginInfo?: any;
  userData?: any;
}

export interface PlayerOptions {
  guildId: string;
  voiceId: string;
  textId?: string;
  deaf?: boolean;
  mute?: boolean;
  volume?: number;
}

export interface FilterConfig {
  volume?: number;
  equalizer?: Array<{
    band: number;
    gain: number;
  }>;
  karaoke?: {
    level?: number;
    monoLevel?: number;
    filterBand?: number;
    filterWidth?: number;
  };
  timescale?: {
    speed?: number;
    pitch?: number;
    rate?: number;
  };
  tremolo?: {
    frequency?: number;
    depth?: number;
  };
  vibrato?: {
    frequency?: number;
    depth?: number;
  };
  rotation?: {
    rotationHz?: number;
  };
  distortion?: {
    sinOffset?: number;
    sinScale?: number;
    cosOffset?: number;
    cosScale?: number;
    tanOffset?: number;
    tanScale?: number;
    offset?: number;
    scale?: number;
  };
  channelMix?: {
    leftToLeft?: number;
    leftToRight?: number;
    rightToLeft?: number;
    rightToRight?: number;
  };
  lowPass?: {
    smoothing?: number;
  };
}